requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
coverage==6.4.1
Django>=3.0,<4.0
psycopg2>=2.8
django-coverage==1.2.4
django-nose==1.4.7
djangorestframework==3.13.1
nose==1.3.7
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
